# Use sync.Pool for parameters