… trie_test.go