# Description
Improve performance with sync.Pool.

# Changes
Pool was used within the parameter process.


Before
```sh
go test -bench=. -cpu=1 -benchmem -count=1
goos: darwin
goarch: amd64
pkg: github.com/bmf-san/goblin
cpu: VirtualApple @ 2.50GHz
BenchmarkStatic1        10446769               119.3 ns/op            32 B/op          1 allocs/op
BenchmarkStatic5         4695543               254.4 ns/op            96 B/op          1 allocs/op
BenchmarkStatic10        2778675               426.5 ns/op           176 B/op          1 allocs/op
BenchmarkWildCard1       1882045               634.5 ns/op           384 B/op          6 allocs/op
BenchmarkWildCard5        497545              2416 ns/op             928 B/op         13 allocs/op
BenchmarkWildCard10       274821              4562 ns/op            1560 B/op         19 allocs/op
BenchmarkRegexp1         1933381               608.6 ns/op           384 B/op          6 allocs/op
BenchmarkRegexp5          515522              2297 ns/op             928 B/op         13 allocs/op
BenchmarkRegexp10         267710              4228 ns/op            1560 B/op         19 allocs/op
PASS
ok      github.com/bmf-san/goblin       13.314s
```

After
```sh
go test -bench=. -cpu=1 -benchmem -count=1
goos: darwin
goarch: amd64
pkg: github.com/bmf-san/goblin
cpu: VirtualApple @ 2.50GHz
BenchmarkStatic1        10187845               118.0 ns/op            32 B/op          1 allocs/op
BenchmarkStatic5         4564369               258.8 ns/op            96 B/op          1 allocs/op
BenchmarkStatic10        2813601               425.1 ns/op           176 B/op          1 allocs/op
BenchmarkWildCard1       1859475               637.9 ns/op           352 B/op          5 allocs/op
BenchmarkWildCard5        522106              2313 ns/op             560 B/op          9 allocs/op
BenchmarkWildCard10       278173              4205 ns/op             802 B/op         14 allocs/op
BenchmarkRegexp1         2031054               606.6 ns/op           352 B/op          5 allocs/op
BenchmarkRegexp5          558216              2180 ns/op             559 B/op          9 allocs/op
BenchmarkRegexp10         304686              3927 ns/op             807 B/op         14 allocs/op
PASS
ok      github.com/bmf-san/goblin       13.436s
```


# Impact range
No specification changes.

# Operational Requirements
N/A

# Related Issue
N/A

# Supplement
N/A